requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.5
pydantic>=1.8.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-dotenv>=0.19.0
numpy>=1.22.0
pytest>=7.0.0
pytest-cov>=4.0.0
httpx>=0.23.0
starlette-csrf>=3.0.0
SQLAlchemy==2.0.25
SpeechRecognition==3.10.3
pydub==0.25.1
astroplan>=0.9
scipy>=1.9.0
pandas>=1.5.0
tqdm>=4.64.0
antares-client>=1.0.0
scikit-learn>=1.1.0
photutils>=1.8.0
matplotlib>=3.5.0
astropy>=5.0.0 